+++ src/Bullet3Common/b3HashMap.h
@@ -43,7 +43,7 @@ struct b3HashString
 		int len = (int)m_string.length();
 		for (int i = 0; i < len; i++)
 		{
-			hash = hash ^ (m_string[(size_t)i]); /* xor  the low 8 bits */
+			hash = hash ^ (unsigned int)(m_string[(size_t)i]); /* xor  the low 8 bits */
 			hash = hash * FNVMultiple;   /* multiply by the magic number */
 		}
 		m_hash = hash;
@@ -73,7 +73,7 @@ struct b3HashString
 	}
 };
 
-const int B3_HASH_NULL = 0xffffffff;
+const int B3_HASH_NULL = (int)0xffffffff;
 
 class b3HashInt
 {
@@ -185,7 +185,7 @@ public:
 		key ^= (key >> 6);
 		key += ~(key << 11);
 		key ^= (key >> 16);
-		return key;
+		return (unsigned int)key;
 	}
 };
 
@@ -219,7 +219,7 @@ public:
 		key ^= (key >> 6);
 		key += ~(key << 11);
 		key ^= (key >> 16);
-		return key;
+		return (unsigned int)key;
 	}
 };
 
@@ -263,7 +263,7 @@ protected:
 				//const Value& value = m_valueArray[i];
 				//const Key& key = m_keyArray[i];
 
-				int hashValue = (int)(m_keyArray[i].getHash() & (m_valueArray.capacity() - 1));  // New hash value with new mask
+				int hashValue = (int)m_keyArray[i].getHash() & (int)(m_valueArray.capacity() - 1);  // New hash value with new mask
 				m_next[i] = m_hashTable[hashValue];
 				m_hashTable[hashValue] = i;
 			}
@@ -273,7 +273,7 @@ protected:
 public:
 	void insert(const Key& key, const Value& value)
 	{
-		int hash = (int)(key.getHash() & (m_valueArray.capacity() - 1));
+		int hash = (int)key.getHash() & (int)(m_valueArray.capacity() - 1);
 
 		//replace value if the key is already there
 		int index = findIndex(key);
@@ -293,7 +293,7 @@ public:
 		{
 			growTables(key);
 			//hash with new capacity
-			hash = (int)(key.getHash() & (m_valueArray.capacity() - 1));
+			hash = (int)key.getHash() & (int)(m_valueArray.capacity() - 1);
 		}
 		m_next[count] = m_hashTable[hash];
 		m_hashTable[hash] = count;
@@ -301,7 +301,7 @@ public:
 
 	void remove(const Key& key)
 	{
-		int hash = (int)(key.getHash() & (m_valueArray.capacity() - 1));
+		int hash = (int)key.getHash() & (int)(m_valueArray.capacity() - 1);
 
 		int pairIndex = findIndex(key);
 
@@ -438,7 +438,7 @@ public:
 
 	int findIndex(const Key& key) const
 	{
-		unsigned int hash = (unsigned int)(key.getHash() & (m_valueArray.capacity() - 1));
+		unsigned int hash = (unsigned int)key.getHash() & (unsigned int)(m_valueArray.capacity() - 1);
 
 		if (hash >= (unsigned int)m_hashTable.size())
 		{
